move default_config_questions to config_questions_helper
